Replace hard-coded minimum hamming scripts with flexible script
This commit adds a script that takes the minimum hamming distance for
dada() as a command-line argument. This new script also combines DADA2
and decontam steps and produces a plot of mock community composition to
help identify a reasonable minimum hamming distance threshold given that
the default setting (i.e., MIN_HAM=1) produces spurious ASVs in the mock
samples.

A batch script is included for exploring multiple minimum hamming
distances in parallel on a SLURM workload manager.