Change trim position to manual setting
This commmit changes how the the forward and reverse read trimming
positions are set. Previously, FIGARO was used to determine the optimal
trim positions. However, the optimal trim positions resulted in too many
ASVs being inferred. In this commit, the trim positions are hard-coded
to cycle 250 for both the forward and reverse reads, because the
inferred mocks have far fewer spurious ASVs. ASVs are now inferred using
the default MINIMUM_HAMMING threshold of 1.